% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_Neff_cal.r
\name{Cell_Neff_cal}
\alias{Cell_Neff_cal}
\title{MERCI DNA rank (Neff counts) calculation}
\usage{
Cell_Neff_cal(
  varMatrix,
  MT_variants,
  MT_coverage,
  donor_cells,
  mixed_cells,
  Ref_nonReceivers = NULL,
  mutFeatures,
  adjust = TRUE,
  adjust.method = "Roe",
  frac_tran = FALSE,
  e = 0.001
)
}
\arguments{
\item{varMatrix}{the processed vaf matrix of mtSNV generated by MTmutMatrix_transform function.}

\item{MT_variants}{the original mtSNV table as the output file generated by readMTvar_10x or readMTvar function.}

\item{MT_coverage}{matrix of Mitochondrial coverage inforamtion (row indicates MT genome locus and column indicates cell), such as the variables generated by readCoverage_10x or readCoverage function.}

\item{donor_cells}{the list of cell names for mitochondrial donor cells.}

\item{mixed_cells}{the list of cell names for candidate receiver cells, they should be mixed cells with unknown ratio of receirves vs non-receivers.}

\item{Ref_nonReceivers}{Only for regular MERCI pipeline, the list of cell names for reference non-receivers. IF not provied, the default is NULL, which means MERCI LOO pipeline will be used.}

\item{mutFeatures}{the list of mutation id for donor cell enriched mtSNVs, which can be obtained from Denrich_mtMut_extr function.}

\item{adjust}{whether adjusting the count statistics into Neff (No.effective count) statistics. when adjust=FALSE, then the original count of observed donor-enriched mtSNV in each candidate receiver cell will be used to calculate the DNA rank. Default=FALSE.}

\item{adjust.method}{Denotes which method to use when adjust=TRUE, Available options are: 'Roe' weight will be calculated as observed(mi)/expected(mi); 'LR' weight will be calculated as the ratio of marinal likelihoods Pi/Qi. When adjust=FALSE, this parameter will be ignored. Default='Roe'.}

\item{frac_tran}{whether using the fraction rather than count/Neff to calculate the DNA rank. Default=FALSE}

\item{e}{sequencing error, default=0.001.}
}
\value{
a data.frame with count/Neff statistics for all mixed_cells and the transformed DNA ranks.
}
\description{
This function calculated the MERCI DNA rank based on the counts of identified donor cell enriched mtSNVs in each input cell.
}
\examples{
Cell_Neff_cal(varMatrix, MT_variants, MT_coverage, donor_cells, mixed_cells, mutFeatures)
}
