% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellNumber_test.r
\name{CellNumber_test}
\alias{CellNumber_test}
\title{Significance estimation for MERCI positive calls}
\usage{
CellNumber_test(
  MTvar_stat,
  MT_constitute_stat,
  top_ranks = seq(100, 10, by = -10),
  Number_R = 1000
)
}
\arguments{
\item{MTvar_stat}{the output of MERCI_MTvar_cal function, it contains the DNA rank score for input mixed cells.}

\item{MT_constitute_stat, }{the output of MERCI_LOO_MT_est or MERCI_MT_est function. it contains the RNA rank scores for input mixed cells.}

\item{top_ranks}{a list of rank cutoffs (top rank \%), default is a list contains 10 cutoffs from top 100\% to top 10\%.}

\item{Number_R}{The times of randomly permutation, default=1000.}
}
\value{
a barlpot showing the Rcm values at different rank cutoffs, and a data.frame with rank cutoffs as rows, associated statistics as columns (p_value, fdr, preN, random_medianN, random_maxN, random_minN, Rcm)
}
\description{
This function tests if enough true-receivers are included in the input mixed cells based on the statistic Rcm.
}
\examples{
CellNumber_test(MTvar_stat, MT_constitute_stat, Number_R=1000) ;
}
