% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_plot.r
\name{coverage_plot}
\alias{coverage_plot}
\title{Plot coverage distribution}
\usage{
coverage_plot(cells, coverage)
}
\arguments{
\item{cells}{a list of cell names}

\item{coverage}{the mitochondrial read coverage information of cells}
}
\value{
a boxplot showing the coverage distribution of selected cells
}
\description{
This function generate a visualization of coverage distrubution for selected cells
}
\examples{
coverage_plot(cells, coverage)
}
