% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Denrich_mtMut_extr.r
\name{Denrich_mtMut_extr}
\alias{Denrich_mtMut_extr}
\title{Identification of donor cell-enriched mtSNVs}
\usage{
Denrich_mtMut_extr(
  varMatrix,
  donor_cells,
  mixed_cells,
  Ref_nonReceivers = NULL,
  OR = 2,
  Nmut_min = 5,
  pvalue = 1,
  qvalue = 1
)
}
\arguments{
\item{varMatrix}{the processed vaf matrix of mtSNV generated by MTmutMatrix_refined function.}

\item{donor_cells}{the list of cell names for mitochondrial donor cells.}

\item{mixed_cells}{the list of cell names for candidate receiver cells, they should be mixed cells with unknown ratio of receirves vs non-receivers.}

\item{Ref_nonReceivers}{Only for regular MERCI pipeline, the list of cell names for reference non-receivers. If provided, then the donor cell-enriched mtSNVs will be obtained by comparing mtSNVs in donor_cells with Ref_nonReceivers, otherwise Ref_nonReceivers will be set as NULL and all mixed_cells will be used as the reference non-recievers to get the donor cell-enriched mtSNVs.}

\item{OR}{Only return the mtSNVs with odds ratio (donor cells vs reference non-reciever cells) > OR, default=2.}

\item{Nmut_min}{the minimum requirement for mutated cell counts, only the mtSNVs with mutated cell counts > Nmut_min will be further used to test if it is a donor cell enriched mtSNV.}

\item{pvalue}{Only return the mtSNVs with statistical p <= pvalue (chi-squre test), defualt=1.}

\item{qvalue}{Only return the mtSNVs with ajusted p value from BH method <= qvalue, default=1.}
}
\description{
This function will identify the a list mitochondrial variants that are more frequently observed in donor cells
}
\examples{
Denrich_mtMut_extr(varMatrix, donor_cells, mixed_cells, OR=2)
}
