% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Donor_enriched_vars.r
\name{Donor_enriched_vars}
\alias{Donor_enriched_vars}
\title{Identification of donor cell enriched mtSNVs}
\usage{
Donor_enriched_vars(
  mutMatrix_processed,
  donor_cells,
  reciever_cells,
  Nmut_min = 5,
  pvalue = 0.05,
  qvalue = 0.1,
  OR = 1
)
}
\arguments{
\item{mutMatrix_processed}{a preprocessed variant VAF matrix with rows are mtSNVs, columns are cells.}

\item{donor_cells}{the list of cell names for mitochondrial donor cells.}

\item{reciever_cells}{the list of cell names for supposed receiver cells, they can be pure reference non-receivers (regular MERCI) or mixed cells with unknown ratio of receirves vs non-receivers (MERCI_LOO).}

\item{Nmut_min}{the minimum requirement for mutated cell counts, only the mtSNVs with mutated cell counts > Nmut_min will be further calculated to determine if it is a donor cell enriched mtSNV.}

\item{pvalue}{Only return the mtSNVs with statistical p <= pvalue, defualt=0.05.}

\item{qvalue}{Only return the mtSNVs with ajusted p value from BH method <= qvalue, default=0.1.}

\item{OR}{Only return the mtSNVs with odds ratio (T cells vs reciever_cells) > OR, default=1.}
}
\value{
a list mtSNVs that is considered as donor cell enriched
}
\description{
This function identifys the donor cell enriched mtSNVs
}
\examples{
Donor_enriched_vars(mutMatrix_processed, donor_cells, reciever_cells, pvalue=0.05, qvalue=0.1, OR=1)
}
