% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErrorMut_filter.r
\name{ErrorMut_filter}
\alias{ErrorMut_filter}
\title{Detect the potential sequencing error}
\usage{
ErrorMut_filter(mut, e = 0.001/3, FPR = 5e-04)
}
\arguments{
\item{mut}{a SNV
e the random sequencing error, default=0.001/3
FPR False positive rate, default=0.0005}
}
\value{
a value of 'detected' or 'error'
}
\description{
This function allows stratify each variant as error or non-error(detected) based on FPR
}
\examples{
ErrorMut_filter(mut, FPR=0.0005)
}
