% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErrorMutmatrix_filter.r
\name{ErrorMutmatrix_filter}
\alias{ErrorMutmatrix_filter}
\title{estimate the potential sequencing error for a variant matrix}
\usage{
ErrorMutmatrix_filter(MT_matrix, FPR = 5e-04)
}
\arguments{
\item{MT_matrix}{a variant matrix (data.frame)
FPR False positive rate, default=0.0005}
}
\value{
a data.frame with the final column showing the estimated resutlts
}
\description{
This function allows stratify each variant in the input matrix as error or non-error(detected) based on FPR
}
\examples{
ErrorMutmatrix_filter(MT_matrix, FPR=0.0005)
}
