% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_MTmuts.r
\name{filter_MTmuts}
\alias{filter_MTmuts}
\title{Preprocess of mtSNV matrix}
\usage{
filter_MTmuts(
  varMatrix = mutMatrix_Hcover,
  MT_coverage,
  donor_cells,
  reciever_cells,
  min_d = 5,
  min_observeRate = 0.1
)
}
\arguments{
\item{varMatrix}{a transformed variant VAF matrix with rows are mtSNVs, columns are cells.}

\item{MT_coverage}{The coverage information of mitochondrial genome for cells.}

\item{donor_cells}{the list of cell names for mitochondrial donor cells.}

\item{reciever_cells}{the list of cell names for candidate receipt cells of MT.}

\item{min_d}{the minimum read coverage required for each mtSNV in different cells, the mutation record will be considered as 'unobservable' and set as NA if the the coverage at the mutation locus < min_d, default=5.}

\item{min_observeRate}{the minimum requirement for observatiion Rate (see parameter min_d), eahc mtSNV will be removed if its observeRate < min_observeRate, default=10\%.}
}
\value{
a processed VAF matrix
}
\description{
This function remove the SNVs with observaRate less than a pre-defined cutoff
}
\examples{
filter_MTmuts(varMatrix, MT_coverage, donor_cells, reciever_cells)
}
