% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MERCI_MT_est.r
\name{MERCI_MT_est}
\alias{MERCI_MT_est}
\title{Decovolution to estimate MT constitute with reference non-receiver data}
\usage{
MERCI_MT_est(
  cell_exp,
  mixed_cells,
  donor_cells,
  Ref_nonReceivers,
  organism = "Human",
  s.markers = NULL,
  epsilon = 0.1
)
}
\arguments{
\item{cell_exp}{the gene expression profiles with rows are gene sumbols and columns are cells, the MT genes must be included.}

\item{mixed_cells}{the list of cell names for input candidates of receiver cells, they should be mixed cells with unknown ratio of receirves vs non-receivers.}

\item{donor_cells}{the list of cell names for mitochondrial donor cells.}

\item{Ref_nonReceivers}{the list of cell names for reference non-receivers.}

\item{organism}{the species of the analyzed sample data,  only supports 'Human' and 'mouse', default='Human'. If s.markers are provided, this parameter can be set as NULL.}

\item{s.markers}{the list of MT genes used for deconvolution, if not provided, then all the MT genes included in the cell_exp will be used, default=NULL.}

\item{epsilon}{epsilon in the insensitive-loss function of ε-SVR, default=0.1.}
}
\value{
a a data.frame with the estimated relative abundance of mitochondria from donor and receiver itself, and the RNA ranks.
}
\description{
This function performas SVR to estiamte the relative abundance for donor-transfered and intrinsic mitochondria.
}
\examples{
MERCI_MT_est(cell_exp, mixed_cells, donor_cells, Ref_nonReceivers, organism='mouse') ;
}
