% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MERCI_MTvar_cal.r
\name{MERCI_MTvar_cal}
\alias{MERCI_MTvar_cal}
\title{DNA rank calculation}
\usage{
MERCI_MTvar_cal(
  varMatrix = mutMatrix_Hcover,
  MT_coverage,
  donor_cells,
  mixed_cells,
  Ref_nonReceivers = NULL,
  min_d = 5,
  min_observeRate = 0.1,
  Nmut_min = 5,
  pvalue = 0.05,
  qvalue = 1,
  OR = 1
)
}
\arguments{
\item{varMatrix}{the processed vaf matrix of mtSNV generated by MTmutMatrix_refined function.}

\item{MT_coverage}{Mitochondrial coverage inforamtion, such as the variables generated by readCoverage_10x or readCoverage function.}

\item{donor_cells}{the list of cell names for mitochondrial donor cells.}

\item{mixed_cells}{the list of cell names for supposed receiver cells, they should be mixed cells with unknown ratio of receirves vs non-receivers.}

\item{min_d}{the minimum read coverage required for each mtSNV in different cells, the mutation record will be considered as 'unobservable' and set as NA if the the coverage at the mutation locus < min_d, default=5.}

\item{min_observeRate}{the minimum requirement for observatiion Rate (see parameter min_d), eahc mtSNV will be removed if its observeRate < min_observeRate, default=10\%.}

\item{Nmut_min}{the minimum requirement for mutated cell counts, only the mtSNVs with mutated cell counts > Nmut_min will be further used to test if it is a donor cell enriched mtSNV.}

\item{pvalue}{Only return the mtSNVs with statistical p <= pvalue (chi-squre test), defualt=0.05.}

\item{qvalue}{Only return the mtSNVs with ajusted p value from BH method <= qvalue, default=0.1.}

\item{OR}{Only return the mtSNVs with odds ratio (T cells vs reciever_cells) > OR, default=1.}
}
\value{
a data.frame with counts of donor cell enriched mtSNVs for all mixed_cells and the DNA ranks.
}
\description{
This function identifys the donor cell enriched mtSNVs and calculated the MERCI DNA rank based on the counts of identified mtSNVs in each input cell.
}
\examples{
MERCI_MTvar_cal(varMatrix, MT_coverage, donor_cells, mixed_cells)
}
