% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MERCI_ReceiverPre.r
\name{MERCI_ReceiverPre}
\alias{MERCI_ReceiverPre}
\title{Predict receivers}
\usage{
MERCI_ReceiverPre(MTvar_stat, MT_constitute_stat, top_rank = 50)
}
\arguments{
\item{MTvar_stat}{the output of MERCI_MTvar_cal function, it contains the DNA rank score for input mixed cells.}

\item{MT_constitute_stat}{the output of MERCI_LOO_MT_est or MERCI_MT_est function. it contains the RNA rank scores for input mixed cells.}

\item{The}{rank cutoff (top rank \%), the cells with both DNA and RNA rank > this cutoff will be considered as receivers, default=50}
}
\value{
A data.frame with input mixed cells as rows, cell_names and prediction labels as columns.
}
\description{
This function predicts the receivers based on adding a 'AND' gate to DNA and RNA rank.
}
\examples{
MERCI_ReceiverPre(MTvar_stat, MT_constitute_stat, top_rank=50) ;
}
