% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTmutMatrix_refined.r
\name{MTmutMatrix_refined}
\alias{MTmutMatrix_refined}
\title{Generate the vaf matrix}
\usage{
MTmutMatrix_refined(MT_variants, coverage)
}
\arguments{
\item{MT_variants}{the mtSNV table with format the same as variables generated by readMTvar_10x or readMTvar function.}

\item{coverage}{the mitochondrial coverage information with content the same as variables generated by readCoverage_10x or readCoverage function.}
}
\value{
a processed vaf matrix, if the locus of mtSNV have no covered read in one sample, the corresponding value was set as NA.
}
\description{
This function generates a vaf matrix with rows as mtSNV and columns as cells
}
\examples{
MTmutMatrix_refined(MT_variants, coverage)
}
