% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTmutMatrix_transform.r
\name{MTmutMatrix_transform}
\alias{MTmutMatrix_transform}
\title{Preprocess the mtSNV table and generate the vaf matrix}
\usage{
MTmutMatrix_transform(
  MT_variants,
  MT_coverage,
  donor_cells,
  mixed_cells,
  Ref_nonReceivers = NULL,
  min_d = 5,
  min_observeRate = 0.1
)
}
\arguments{
\item{MT_variants}{the mtSNV table with format the same as variables generated by readMTvar_10x or readMTvar function.}

\item{MT_coverage}{the mitochondrial coverage information with content the same as variables generated by readCoverage_10x or readCoverage function.}

\item{donor_cells}{the list of cell names for mitochondrial donor cells.}

\item{mixed_cells}{the list of cell names for candidate receiver cells, they should be mixed cells with unknown ratio of receirves vs non-receivers.}

\item{Ref_nonReceivers}{Only for regular MERCI pipeline, the list of cell names for reference non-receivers. IF not provied, the default is NULL, which means MERCI LOO pipeline will be used.}

\item{min_d}{the minimum read coverage required for each mtSNV in different cells, the mutation record will be considered as 'unobservable' and set as NA if the the coverage at the mutation locus < min_d, default=5.}

\item{min_observeRate}{the minimum requirement for observatiion Rate (see parameter min_d), any mtSNV with observeRate < min_observeRate will be removed, default=10\%.}
}
\value{
a processed vaf matrix, if the locus of mtSNV have no covered read or read number < min_d in one sample, the corresponding value was set as NA.
}
\description{
This function generates a vaf matrix with rows as mtSNV and columns as cells
}
\examples{
MTmutMatrix_transform(MT_variants, MT_coverage, donor_cells, mixed_cells)
}
