% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mut_inf_extr.r
\name{mut_inf_extr}
\alias{mut_inf_extr}
\title{mutation annotation for a specified mutation list}
\usage{
mut_inf_extr(
  muts,
  donor_cell_ref,
  receiver_cell_ref,
  mutMatrix_processed,
  MT_variants,
  MT_coverage
)
}
\arguments{
\item{muts}{the mitochondrial mutation list.}

\item{donor_cell_ref}{reference donor cells.}

\item{mutMatrix_processed}{the processed vaf matrix of mtSNV generated by MTmutMatrix_refined function.}

\item{MT_variants}{the mtSNV table with format the same as variables generated by readMTvar_10x or readMTvar function.}

\item{MT_coverage}{the mitochondrial coverage information with content the same as variables generated by readCoverage_10x or readCoverage function.}
}
\value{
a data.frame with annotation of mutation list muts.
}
\description{
This function extracted the detailed mutation annotation.
}
\examples{
mut_inf_extr(muts, donor_cell_ref, receiver_cell_ref, mutMatrix_processed, MT_variants, MT_coverage)
}
