% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Random_perfs_test.r
\name{Random_perfs_test}
\alias{Random_perfs_test}
\title{The expected postive cell numbers by add 'AND' gate to random ranks}
\usage{
Random_perfs_test(MTreceiver_pre, top_rank = 20, Number_R = 10)
}
\arguments{
\item{MTreceiver_pre}{the predicted results of MERCI, which a the file generated by Integ_rank_pre function.}

\item{top_rank}{the top percentage of rank scores, the positive calls will generated based on this cutoff, default=20.}

\item{Number_R}{the times of randomly permutation, default=10.}
}
\value{
a dataframe with the number of positive calls of MERCI and random.
}
\description{
This function will returen the postive calls from random ranks.
}
\examples{
Random_perfs_test(Treceiver_pre, top_rank=50, Number_R=1000)
}
