% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCoverage.r
\name{readCoverage}
\alias{readCoverage}
\title{Read in the mitochondrial read coverage inforamtion}
\usage{
readCoverage(CoverFile)
}
\arguments{
\item{CoverFile}{The path to .MT_Coverage.csv file generated by MERCI-mtSNP.}
}
\value{
a data.frame of MT coverage inforamtion
}
\description{
This function read in MT coverage inforamtion (*.MT_Coverage.csv) generated by MERCI-mtSNP, works for data types: smart-seq2, bulk RNA-seq, scATAC-seq, bulk ATAC-seq.
}
\examples{
readCoverage(CoverFile='./MT_Coverage.csv')
}
