% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCoverage_10x.r
\name{readCoverage_10x}
\alias{readCoverage_10x}
\title{Read in the mitochondrial read coverage inforamtion}
\usage{
readCoverage_10x(CoverFile, S.cells = NULL)
}
\arguments{
\item{CoverFile}{The path to .Coverage_Cell.csv file generated by MERCI-mtSNP.}

\item{S.cells}{a list of cells, the MT coverage of only these selected cells will be returned. if not provided then all cells will be considered, default=NULL.}
}
\value{
a boxplot showing the MT coverage distribution across cells, a data.frame for MT coverage inforamtion
}
\description{
This function read in MT coverage inforamtion (*.Coverage_Cell.csv) generated by MERCI-mtSNP, only works for 10x_scRNA-seq data type.
}
\examples{
readCoverage_10x(CoverFile='./X.Coverage_Cell.csv', S.cells=NULL)
}
