% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMTvar.r
\name{readMTvar}
\alias{readMTvar}
\title{read in mtSNV of single cell sequencing data}
\usage{
readMTvar(varFile, cellname = "XXX")
}
\arguments{
\item{varFile}{The path to .MT_variants.txt file generated by MERCI-mtSNP.}

\item{cellname}{A character reprents the cell indentifier.}
}
\value{
the mtSNV annotation table
}
\description{
This function read in MT variants table (*.MT_variants.txt) generated by MERCI-mtSNP, works for data types: smart-seq2, bulk RNA-seq, scATAC-seq, bulk ATAC-seq.
}
\examples{
readMTvar(varFile='./X.MT_variants.txt', cellname='XXX')
}
