% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMTvar_10x.r
\name{readMTvar_10x}
\alias{readMTvar_10x}
\title{read in mtSNV of 10x scRNA-seq data}
\usage{
readMTvar_10x(varFile, minReads = 500)
}
\arguments{
\item{varFile}{The path to .MT_variants.txt file generated by MERCI-mtSNP.}

\item{minReads}{Only mtSNVs in Cells with MT read counts > minReads will be retained.}
}
\value{
the mtSNV annotation table for cells with MT read counts more than a user-defined number
}
\description{
This function read in MT variants table (*.MT_variants.txt) generated by MERCI-mtSNP, only works for 10x_scRNA-seq data type.
}
\examples{
readMTvar_10x(varFile='./X.MT_variants.txt', minReads=500)
}
