% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_cal.r
\name{w_cal}
\alias{w_cal}
\title{The weights (obs(mi)/exp(mi) ratios) MT mutations}
\usage{
w_cal(muts, t_cell, muts_inf, mutMatrix_processed, MT_coverage, e = 0.001)
}
\arguments{
\item{muts}{the mitochondrial mutation list.}

\item{t_cell}{a cell name or barcode.}

\item{muts_inf}{the annotation of muts, also as the output of mut_inf_extr function.}

\item{mutMatrix_processed}{the processed vaf matrix of mtSNV generated by MTmutMatrix_refined function.}

\item{e}{sequencing error, default=0.001.}
}
\value{
a a list weights with same length of muts.
}
\description{
This function calculates the weights of MT mutations in a given cell.
}
\examples{
w_cal(muts, t_cell, muts_inf, mutMatrix_processed)
}
